requirements.txt

Django
twilio
whitenoise
gunicorn
pytz
Pillow
django-adminlte-ui
djangorestframework